+++ TestTask2/Models/Product.cs
@@ -13,7 +13,7 @@ namespace TestTask2.Models
 
         public virtual ICollection<Image> Images { get; set; }
 
-        public Product(string domain, string descrption, int price, ICollection<Image> images, Currency currency) : base(domain, descrption, price, currency)
+        public Product(string domain, string descrption, decimal price, ICollection<Image> images, Currency currency) : base(domain, descrption, price, currency)
         {
             this.Images = images;
         }
